エラーメッセージ表示の修正
