--- src/components/Site/EditSiteInfo/EditSiteInfo.tsx
@@ -1,8 +1,0 @@
-
-const EditSiteInfo = () => {
-  return (
-    <div>EditSiteInfo</div>
-  )
-}
-
-export default EditSiteInfo
